Claves foraneas ready
--> La tabla conductroes tiene clave foranea a la tabla de usuarios y a la empresa.
->La tabla empresas tiene clave foranea a la tabla de usuarios.

+++ database/migrations/2019_06_24_155512_create_empresas_table.php
@@ -14,8 +14,10 @@ class CreateEmpresasTable extends Migration
     public function up()
     {
         Schema::create('empresas', function (Blueprint $table) {
+            //el id propio de esta tabla
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id');
+            //con el  que se referencia en la tabla usuarios
+            $table->unsignedBigInteger('user_id');
             $table->string('nit')->unique();
             $table->string('representante-legal'); //nombre de representante legal.
             $table->timestamps();
